Move common structs to module_versioning.go and module_set_release.go files
Exported and moved RepoRootTag to conversions.go, to be used in future tag tests
Updated testDataDir to use absolute path rather than relative
