+++ multimod/internal/common/conversions.go
@@ -19,13 +19,17 @@ import (
 	"strings"
 )
 
+const (
+	RepoRootTag = ModuleTagName("REPOROOTTAG")
+)
+
 // combineModuleTagNamesAndVersion combines a slice of ModuleTagNames with the version number and returns
 // the new full module tags.
 func combineModuleTagNamesAndVersion(modTagNames []ModuleTagName, version string) []string {
 	var modFullTags []string
 	for _, modTagName := range modTagNames {
 		var newFullTag string
-		if modTagName == repoRootTag {
+		if modTagName == RepoRootTag {
 			newFullTag = version
 		} else {
 			newFullTag = string(modTagName) + "/" + version
@@ -80,7 +84,7 @@ func moduleFilePathToTagName(modFilePath ModuleFilePath, repoRoot string) (Modul
 
 	// if the modTagName is equal to go.mod, it is the root repo
 	if modTagName == "go.mod" {
-		return repoRootTag, nil
+		return RepoRootTag, nil
 	}
 
 	return ModuleTagName(modTagName), nil
