+++ multimod/internal/common/test_data/verify_git_tags_do_not_already_exist/versions_valid.yaml
@@ -17,10 +17,14 @@ module-sets:
     version: v1.2.3-RC1+meta
     modules:
       - go.opentelemetry.io/test/test1
-      - go.opentelemetry.io/test/test2
+      - go.opentelemetry.io/test/test4
   mod-set-2:
     version: v0.1.0
     modules:
       - go.opentelemetry.io/test3
+  mod-set-3:
+    version: v2.2.2
+    modules:
+      - go.opentelemetry.io/testroot/v2
 excluded-modules:
-  - go.opentelemetry.io/excluded1
+  - go.opentelemetry.io/test/testexcluded
